feat: search-card redesign
- share button
- copy link to clipboard button
