+++ public/assets/entry.js
@@ -1,1 +1,1 @@
-const r=/^https:\/\/(open\.spotify\.com\/(track|album|playlist|artist|episode|show)|spotify\.link)\/(\w{11,24})(?:[\?#].*)?$/,n=/(?:https?:\/\/)?(?:www\.)?(?:youtube\.com|youtu\.be|music\.youtube\.com)\/(?:watch\?v=|embed\/|v\/|shorts\/|playlist\?list=|channel\/)?([\w-]{11,})(?:\S+)?(?:&si=\S+)?/,a=new URLSearchParams(window.location.search),c=({link:e})=>{const o=document.getElementById("search-form");o.querySelector("input").value=e,htmx.ajax("POST","/search",{source:"#search-form"})},s=async()=>{if("clipboard"in navigator)try{const e=await navigator.clipboard.readText();e.match(new RegExp(`${r.source}|${n.source}`))&&!a.get("id")&&c({link:e})}catch(e){console.error("Clipboard access error:",e)}else console.error("Clipboard API is not supported.")};document.addEventListener("htmx:afterOnLoad",()=>{const e=document.getElementById("search-card")?.getAttribute("data-id");e&&(a.set("id",e),window.history.replaceState({},"",`${window.location.pathname}?${a}`))});document.addEventListener("htmx:timeout",function(){document.getElementById("search-results").innerHTML='<p class="mt-8 text-center">Something went wrong, try again later.</p>'});document.addEventListener("DOMContentLoaded",async()=>{await s()});window.shareLink=async e=>{if(!e){console.error("Universal link not found");return}const o=new Notyf({riple:!1,dismissible:!0,duration:2e3,types:[{type:"success",background:"black"}]});try{if(navigator.share){await navigator.share({title:"Share this universal link",url:e});return}if(navigator.clipboard)await navigator.clipboard.writeText(e);else{const t=document.createElement("textarea");t.value=e,document.body.appendChild(t),t.select(),document.execCommand("copy"),document.body.removeChild(t)}o.success("Link copied to clipboard!")}catch(t){console.error(t)}};
+const r=/^https:\/\/(open\.spotify\.com\/(track|album|playlist|artist|episode|show)|spotify\.link)\/(\w{11,24})(?:[\?#].*)?$/,n=/(?:https?:\/\/)?(?:www\.)?(?:youtube\.com|youtu\.be|music\.youtube\.com)\/(?:watch\?v=|embed\/|v\/|shorts\/|playlist\?list=|channel\/)?([\w-]{11,})(?:\S+)?(?:&si=\S+)?/,a=new URLSearchParams(window.location.search),c=({link:e})=>{const t=document.getElementById("search-form");t.querySelector("input").value=e,htmx.ajax("POST","/search",{source:"#search-form"})},i=async()=>{if("clipboard"in navigator)try{const e=await navigator.clipboard.readText();e.match(new RegExp(`${r.source}|${n.source}`))&&!a.get("id")&&c({link:e})}catch(e){console.error("Clipboard access error:",e)}else console.error("Clipboard API is not supported.")};document.addEventListener("htmx:afterOnLoad",()=>{const e=document.getElementById("search-card")?.getAttribute("data-id");e&&(a.set("id",e),window.history.replaceState({},"",`${window.location.pathname}?${a}`))});document.addEventListener("htmx:timeout",function(){document.getElementById("search-results").innerHTML='<p class="mt-8 text-center">Something went wrong, try again later.</p>'});document.addEventListener("DOMContentLoaded",async()=>{await i()});window.shareLink=async e=>{if(!e){console.error("Universal link not found");return}try{if(navigator.share){await navigator.share({title:"Share this universal link",url:e});return}await copyLinkToClipboard(e)}catch(t){console.error(t)}};window.copyLinkToClipboard=async e=>{const t=new Notyf({riple:!1,dismissible:!0,duration:2e3,types:[{type:"success",background:"black"}]});if(navigator.clipboard)await navigator.clipboard.writeText(e);else{const o=document.createElement("textarea");o.value=e,document.body.appendChild(o),o.select(),document.execCommand("copy"),document.body.removeChild(o)}t.success("Link copied to clipboard!")};
