+++ src/views/components/search-card.tsx
@@ -1,40 +1,103 @@
+import { Adapter } from '~/config/enum';
 import { SearchResult } from '~/services/search';
 
-import SearchLink from './search-link';
+const SEARCH_LINK_DICT = {
+  [Adapter.Spotify]: {
+    icon: 'fab fa-spotify',
+    label: 'Listen on Spotify',
+  },
+  [Adapter.YouTube]: {
+    icon: 'fab fa-youtube',
+    label: 'Listen on YouTube Music',
+  },
+  [Adapter.Deezer]: {
+    icon: 'fab fa-deezer',
+    label: 'Listen on Deezer',
+  },
+  [Adapter.AppleMusic]: {
+    icon: 'fab fa-apple',
+    label: 'Listen on Apple Music',
+  },
+  [Adapter.Tidal]: {
+    icon: 'fa fa-music',
+    label: 'Listen on Tidal',
+  },
+  [Adapter.SoundCloud]: {
+    icon: 'fab fa-soundcloud',
+    label: 'Listen on SoundCloud',
+  },
+};
+
+function SearchLink(props: { type: Adapter; url: string; isVerified?: boolean }) {
+  const searchResult = SEARCH_LINK_DICT[props.type];
+
+  return (
+    <li class="flex items-center justify-between gap-1 rounded-lg p-2 text-sm hover:bg-zinc-700 sm:text-base">
+      <a
+        href={props.url}
+        target="_blank"
+        rel="noreferrer"
+        aria-label={searchResult.label}
+        class="flex items-center"
+      >
+        <i class={`${searchResult.icon} w-8`} />
+        <p class="underline decoration-0 underline-offset-2">{searchResult.label}</p>
+        {props.isVerified && (
+          <span
+            class="ml-1 inline-flex items-center justify-center rounded-full bg-green-500 p-1 text-[0.56rem]"
+            aria-label="Verified"
+          >
+            <i class="fas fa-check" />
+          </span>
+        )}
+      </a>
+      <button type="button" onclick={`copyLinkToClipboard('${props.url}')`}>
+        <i class="fa fa-regular fa-copy" />
+      </button>
+    </li>
+  );
+}
 
 export default function SearchCard(props: { searchResult: SearchResult }) {
   return (
     <div
       id="search-card"
       data-id={props.searchResult.id}
-      class="relative m-4 flex max-w-2xl flex-wrap items-start justify-center rounded-lg border border-white md:p-4"
+      class="relative m-4 flex flex-wrap items-start justify-center gap-4 rounded-lg shadow-lg md:p-4"
     >
-      <button
-        type="button"
-        class="absolute right-2 top-2 p-1 text-white"
-        onclick={`shareLink('${props.searchResult.universalLink}')`}
-      >
-        <i class="fas fa-up-right-from-square" />
-      </button>
-      <div class="m-4 w-full md:w-44">
+      <div class="flex w-full items-center justify-start gap-4">
         <img
-          class="mx-auto w-28 md:w-44"
+          class="w-28 rounded-lg"
           src={props.searchResult.image}
           alt={props.searchResult.title}
         />
-      </div>
-      <div class="mb-2 flex-1 flex-col items-start p-2 md:mr-6">
-        <div class="mb-2 hyphens-auto text-center text-2xl font-normal md:text-start">
-          {props.searchResult.title}
+        <div class="flex flex-col gap-1">
+          <h3 class="hyphens-auto text-center text-2xl font-normal md:text-start">
+            {props.searchResult.title}
+          </h3>
+          <p class="text-center text-sm text-zinc-400 md:text-start">
+            {props.searchResult.description}
+          </p>
+          <div class="mt-2 flex">
+            <button
+              type="button"
+              class="flex items-center justify-center gap-2 rounded-lg bg-zinc-700 px-3 py-1 text-sm font-semibold"
+              onclick={`shareLink('${props.searchResult.universalLink}')`}
+            >
+              <i class="fas fa-arrow-up-from-bracket" />
+              Share
+            </button>
+          </div>
         </div>
-        <p class="text-center text-sm md:text-start">{props.searchResult.description}</p>
+      </div>
+      <div class="flex min-h-12 flex-1 flex-col items-start p-2">
         {props.searchResult.links.length === 0 && (
-          <p class="mt-6 w-full text-center text-sm md:text-start">
+          <p class="w-full text-center text-sm md:text-start">
             Not available on other platforms
           </p>
         )}
         {props.searchResult.links.length > 0 && (
-          <ul class="mt-4 min-w-48 text-base">
+          <ul class="w-full">
             {props.searchResult.links.map(({ type, url, isVerified }) => (
               <SearchLink type={type} url={url} isVerified={isVerified} />
             ))}
